move OnlineConfig
